requirements.txt
django==4.2.2
requests
djangorestframework==3.14.0
pyyaml
django-cors-headers==4.0.0
selenium==4.10.0
beautifulsoup4==4.12.2
lxml